# Correctly escape ASCII control characters in strings