--- a/json/src/ser.rs
+++ b/json/src/ser.rs
@@ -504,6 +504,17 @@ pub fn escape_bytes<W>(wr: &mut W, bytes: &[u8]) -> Result<()>
             b'\n' => b"\\n",
             b'\r' => b"\\r",
             b'\t' => b"\\t",
+            c @ b'\x00' ... b'\x1F' => {
+                if start < i {
+                    try!(wr.write_all(&bytes[start..i]));
+                }
+
+                try!(write!(wr,"\\u{:04X}", c));
+
+                start = i + 1;
+
+                continue;
+            },
             _ => { continue; }
         };