This PR fixes #51.

I also took the liberty to add more tests in both string serialization and deserialization since it's very easy to miss some subtle parts of the standard.

The code for escaping control characters is definitely not very elegant but the other solutions I found were either using a vec (with the associated heap allocation and bound checking), having more code duplication or having worst control flow.

I someone suggest a better solution, I'd be happy to amend this PR.