@PaulGrandperrin 2016-04-25
The travis build fails on rust nightly but it doesn't seem to be related to this PR changes.