@dtolnay 2016-04-26
7F and 80-9F are also Unicode control characters.