@PaulGrandperrin 2016-04-26
It's true that theses codepoints are considered as control characters by Unicode itself, but they are not explicitly mentioned in the JSON specification:
http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf

However it would be absolutely legal and probably a good idea to escape them anyway.

Ruby does escape them for instance.

I'll add them.