@PaulGrandperrin 2016-04-30
Now that `ser::escape_str()` has its own implementation, I think we can remove `ser::escape_bytes()` and replace all references to it.

It will however break the API so I don't know if I should proceed.

Maybe we could reimplement `ser::escape_bytes()` using `ser::escape_str()` with [String::from_utf8](https://doc.rust-lang.org/nightly/std/string/struct.String.html#method.from_utf8) or [String::from_utf8_lossy](https://doc.rust-lang.org/nightly/std/string/struct.String.html#method.from_utf8_lossy) or [String::from_utf8_unchecked](https://doc.rust-lang.org/nightly/std/string/struct.String.html#method.from_utf8_unchecked)