@dtolnay 2016-05-01
Nicely done. Thanks for seeing this through. I filed #60 to get rid of `escape_bytes` in the next breaking release.

@oli-obk or @erickt please take a look.