@oli-obk 2016-05-01
The escape_bytes codepaths aren't tested anymore, since they arent called in the string conversion. Add specific tests?