@oli-obk 2016-05-01
Wonderful. I don't understand the algorithm, but the tests lgtm