@PaulGrandperrin 2016-05-01
I also ran the benchmarks and we can see that some tests took a noticeable performance hit:

- master (7bc9b0a)
```
test bench_log::bench_copy                                      ... bench:          35 ns/iter (+/- 2) = 17285 MB/s
test bench_log::bench_decoder                                   ... bench:      18,073 ns/iter (+/- 1,326) = 33 MB/s
test bench_log::bench_deserializer                              ... bench:       5,493 ns/iter (+/- 137) = 110 MB/s
test bench_log::bench_encoder                                   ... bench:       3,336 ns/iter (+/- 56) = 181 MB/s
test bench_log::bench_manual_serialize_my_mem_writer0_escape    ... bench:       3,252 ns/iter (+/- 399) = 186 MB/s
test bench_log::bench_manual_serialize_my_mem_writer0_no_escape ... bench:       2,003 ns/iter (+/- 50) = 302 MB/s
test bench_log::bench_manual_serialize_my_mem_writer1_escape    ... bench:       1,934 ns/iter (+/- 78) = 312 MB/s
test bench_log::bench_manual_serialize_my_mem_writer1_no_escape ... bench:       1,124 ns/iter (+/- 163) = 538 MB/s
test bench_log::bench_manual_serialize_vec_escape               ... bench:       2,052 ns/iter (+/- 163) = 294 MB/s
test bench_log::bench_manual_serialize_vec_no_escape            ... bench:       1,315 ns/iter (+/- 36) = 460 MB/s
test bench_log::bench_serializer                                ... bench:       2,605 ns/iter (+/- 112) = 232 MB/s
test bench_log::bench_serializer_my_mem_writer0                 ... bench:       3,613 ns/iter (+/- 689) = 167 MB/s
test bench_log::bench_serializer_my_mem_writer1                 ... bench:       2,190 ns/iter (+/- 35) = 276 MB/s
test bench_log::bench_serializer_slice                          ... bench:       3,069 ns/iter (+/- 1,791) = 197 MB/s
test bench_log::bench_serializer_vec                            ... bench:       2,369 ns/iter (+/- 352) = 255 MB/s
```

- this PR (632a555)
```
test bench_log::bench_copy                                      ... bench:          35 ns/iter (+/- 2) = 17285 MB/s
test bench_log::bench_decoder                                   ... bench:      18,576 ns/iter (+/- 170) = 32 MB/s
test bench_log::bench_deserializer                              ... bench:       5,594 ns/iter (+/- 905) = 108 MB/s
test bench_log::bench_encoder                                   ... bench:       3,214 ns/iter (+/- 174) = 188 MB/s
test bench_log::bench_manual_serialize_my_mem_writer0_escape    ... bench:       3,692 ns/iter (+/- 310) = 163 MB/s
test bench_log::bench_manual_serialize_my_mem_writer0_no_escape ... bench:       1,998 ns/iter (+/- 196) = 302 MB/s
test bench_log::bench_manual_serialize_my_mem_writer1_escape    ... bench:       2,580 ns/iter (+/- 37) = 234 MB/s
test bench_log::bench_manual_serialize_my_mem_writer1_no_escape ... bench:       1,103 ns/iter (+/- 39) = 548 MB/s
test bench_log::bench_manual_serialize_vec_escape               ... bench:       2,698 ns/iter (+/- 199) = 224 MB/s
test bench_log::bench_manual_serialize_vec_no_escape            ... bench:       1,286 ns/iter (+/- 117) = 470 MB/s
test bench_log::bench_serializer                                ... bench:       3,188 ns/iter (+/- 67) = 189 MB/s
test bench_log::bench_serializer_my_mem_writer0                 ... bench:       4,138 ns/iter (+/- 330) = 146 MB/s
test bench_log::bench_serializer_my_mem_writer1                 ... bench:       2,707 ns/iter (+/- 55) = 223 MB/s
test bench_log::bench_serializer_slice                          ... bench:       3,786 ns/iter (+/- 2,035) = 159 MB/s
test bench_log::bench_serializer_vec                            ... bench:       2,914 ns/iter (+/- 74) = 207 MB/s
```

Here are the most noticeable:
manual_serialize_my_mem_writer1_escape : **75%** of original speed
manual_serialize_vec_escape: **76%** of original speed
serializer: **81%** of original speed
serializer_my_mem_writer1: **81%** of original speed
serializer_slice: **81%** of original speed
serializer_vec: **81%** of original speed

However this new code better reuse UTF-8 code from the stdlib and, as the PR says, correctly escapes control characters.

My CPU: i7-3517U