@PaulGrandperrin 2016-05-01
And BTW, the "newly rewritten but soon to be removed" code of `escape_bytes()` also passes all the hundreds of new tests and has almost exactly the same performance profile as the new `escape_str()`.