@dtolnay 2016-05-01
Thanks for calling out the performance difference but I am not concerned about that. When we have two correct implementations that we need to decide between, we can turn to performance to guide us, but here we are talking about an incorrect implementation vs a correct implementation so performance is secondary.

@oli-obk I think your feedback has been addressed and this is ready to merge.