@oli-obk 2016-05-01
> @oli-obk I think your feedback has been addressed and this is ready to merge.

As I said on the issue, I want to clear the rules on merging with @erickt first

But this looks good to merge to me