@homu 2016-06-25
:umbrella: The latest upstream changes (presumably #85) made this pull request unmergeable. Please resolve the merge conflicts.